pyproject.toml
[project]
name = "pydemetra"
version = "0.1.0"
description = "Python bindings for JDemetra+ seasonal adjustment and time series analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jpype1>=1.5.0",
    "protobuf>=5.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "loguru>=0.7",
    "scipy>=1.10",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pre-commit>=4.5.1",
    "pyarrow>=23.0.1",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.15.4",
    "ty>=0.0.20",
    "xdoctest>=1.1",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pydemetra*"]

[tool.setuptools.package-data]
"pydemetra._jars" = ["*.jar"]
"pydemetra._proto" = ["*.py"]

[tool.ruff]
line-length = 99
exclude = ["pydemetra/_proto/*_pb2.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
